Move named keys related marcos to core
